feat: Add support for base64 and OAuth tokens
